react-agent/pyproject.toml
[project]
name = "ice-breaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix>=5.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.2",
    "langchain-classic>=1.0.0",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.2.1",
]
